Modify embed_tfidf stage to use save_npz
- Saving the trainX, testX and validX sparse matrices (scipy.sparse.csr_matrix) using np.save causes
a problem. So, when we save an object using np.save (e.g. np.save(x)), it first makes a call to turn
that object into an np ndarray (using np.array(x)). In the case of a csr_matrix, or a non-iterable
Python object, this will return a 0-dimensional array, which will cause problems for sklearn funcs.
- There is a way to get back those underlying objects, by calling zero_dim_array.item(). But it's
better to use scipy.sparse.[save/load]_npz, as it stores the sparse array in a compressed format.