Update Dockerfile: 3-stage build and smaller context
- The Dockerfile now has 3 build stages to improve our ability to use the layer compression and cache
mechanism. The first stage downloads and installs the system packages, and will change extremely rarely.
The second stage installs the python packages, and will change more frequently than the first stage,
but still much rarely compared to the last stage. The final stage copies the package code, pulls the
artifacts using DVC, installs the root as a package (using poetry), and then deletes the dvc and git
files.

- The .dockerignore file has also been improved to minimize the build context sent to the docker
build server.

One of the main factors for this change was to be able to use cloud container builds, e.g. on Docker
Hub (or Github, Gitlab, Heroku, AWS).

- The application is now run using gunicorn (ASGI server) in async mode, with default 2 uvicorn
workers and 8 threads per worker.